Backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.3.0
flask-jwt-extended==4.5.3
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
xgboost==1.7.6
python-dotenv==1.0.0
psycopg2-binary==2.9.7
redis==5.0.1
requests==2.31.0
cryptography==41.0.7
prometheus-client==0.17.1
reportlab==4.0.4
websockets==12.0
asyncio==3.4.3
psutil==5.9.5
gunicorn==21.2.0